pyproject.toml
[project]
name = "NumberNinja"
description = "A personal bookkeeping application"
authors = [
  { name="René Moll" },
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
	"Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dynamic = ["version"]

dependencies = [
  "python-dateutil",
  "py-money>=0.5",
]

[project.optional-dependencies]
dev = [
	"pre-commit",
]
test = [
  "pytest>=8.3",
  "pytest-cov>=5.0"
]
behaviour = [
  "behave"
]

[project.urls]
Homepage = "https://github.com/renemoll/numberninja"
Issues = "https://github.com/renemoll/numberninja/issues"

#
# Entry points
#

[project.scripts]
numberninja-cli = "numberninja:cli"

#
# Build system
#

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "numberninja.__version__"}

#
# Linters
#

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "FA",     # FA: flake8-future-annotations
  "COM812", # COM812: trailing comma missing
  "DTZ007", # DTZ007: call-datetime-strptime-without-zone (TODO: remove)
  "E501",   # E501: line too long
  "EM101",  # EM101: raw-string-in-exception
  "TRY003", # TRY003: raise-vanilla-args
]

[tool.ruff.lint.per-file-ignores]
"noxfile.py" = ["ALL"]
"test/features/steps/*" = [
  "ANN",  # ANN: flake8-annotations
  "D",    # D: pydocstyle
  "F811", # F811: redefined-while-unused
  "INP",  # INP: flake8-no-pep420
]
"test/*.py" = [
  "ANN",     # ANN: flake8-annotations
  "D",       # D: pydocstyle
  "PLR2004", # PLR2004: magic-value-comparison
  "S101",    # S101: assert
]

[tool.ruff.lint.pydocstyle]
convention = "google"

#
# Testing
#

[tool.coverage]
html.directory = "build/coverage/"
report.show_missing = true
report.fail_under = 90
run.branch = true
run.source = ["numberninja"]

[tool.pytest.ini_options]
addopts = "--verbose --showlocals --cov"
